+++ src/main/java/io/github/co_mmer/aaamockmvc/test/web/asserts/TestAssertImpl.java
@@ -1,33 +1,31 @@
 package io.github.co_mmer.aaamockmvc.test.web.asserts;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.github.co_mmer.aaamockmvc.test.web.asserts.content.TestAssertContent;
 import io.github.co_mmer.aaamockmvc.test.web.asserts.content.TestAssertContentImpl;
+import io.github.co_mmer.aaamockmvc.test.web.asserts.custom.TestAssertCustom;
+import io.github.co_mmer.aaamockmvc.test.web.asserts.custom.TestAssertCustomImpl;
 import io.github.co_mmer.aaamockmvc.test.web.asserts.head.TestAssertHead;
 import io.github.co_mmer.aaamockmvc.test.web.asserts.head.TestAssertHeadImpl;
+import io.github.co_mmer.aaamockmvc.test.web.asserts.status.TestAssert1Status;
+import io.github.co_mmer.aaamockmvc.test.web.asserts.status.TestAssertStatusImpl;
 import lombok.NonNull;
-import org.junit.jupiter.api.Assertions;
-import org.springframework.http.HttpStatus;
-import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.test.web.servlet.ResultMatcher;
 
 /**
  * This class provides implementation for assertions on HTTP responses in a testing context. It
- * allows various validations of response status, content, and headers.
+ * allows various validations of response status, content, headers, and custom assertions.
  *
- * <p>It is primarily used for validating the results of HTTP requests performed with the {@code
- * MockMvc} framework in a Spring web application context.
+ * <p>It is primarily used to validate the results of HTTP requests performed with the {@code
+ * MockMvc} framework in a Spring web application context. This class provides specific
+ * implementations for asserting the status, content, and headers of an HTTP response, as well as
+ * the option to define custom assertions.
  *
  * @since 1.0.0
  */
 public final class TestAssertImpl implements TestAssert {
 
   private final ResultActions actions;
-  private final MockHttpServletResponse response;
   private final ObjectMapper objectMapper;
 
   /**
@@ -42,54 +40,22 @@ public final class TestAssertImpl implements TestAssert {
    */
   public TestAssertImpl(@NonNull ResultActions actions, @NonNull ObjectMapper objectMapper) {
     this.actions = actions;
-    this.response = actions.andReturn().getResponse();
     this.objectMapper = objectMapper;
   }
 
   /**
-   * Asserts that the status of the HTTP response matches the given {@code HttpStatus}.
+   * Asserts that the HTTP response status matches the expected status code.
    *
-   * @param status the expected {@code HttpStatus} of the response (must not be {@code null})
-   * @return the current instance of {@code TestAssert} for method chaining
-   * @since 1.0.0
+   * <p>This method returns an instance of {@code TestAssertStatus} for asserting the status of the
+   * HTTP response. It allows various validations of response status codes, ensuring that the
+   * response matches the expected outcomes of the HTTP request.
+   *
+   * @return an instance of {@code TestAssertStatus} for further assertions on the response status
+   * @since 1.1.0
    */
   @Override
-  public TestAssert assertStatus(@NonNull HttpStatus status) {
-    assertThat(this.response.getStatus(), is(status.value()));
-    return this;
-  }
-
-  /**
-   * Asserts that the status of the HTTP response matches the given status code.
-   *
-   * @param status the expected status code of the response
-   * @return the current instance of {@code TestAssert} for method chaining
-   * @since 1.0.0
-   */
-  @Override
-  public TestAssert assertStatus(int status) {
-    assertThat(this.response.getStatus(), is(status));
-    return this;
-  }
-
-  /**
-   * Asserts that the result of the test request matches the given {@link ResultMatcher}.
-   *
-   * <p>If an error occurs, execution is terminated with a call to {@code Assertions.fail}, passing
-   * the corresponding exception.
-   *
-   * @param matcher the {@code ResultMatcher} to be used for validation (must not be {@code null})
-   * @return the current instance of {@code TestAssert} for method chaining
-   * @since 1.0.0
-   */
-  @Override
-  public TestAssert assertByResultMatcher(@NonNull ResultMatcher matcher) {
-    try {
-      this.actions.andExpect(matcher);
-    } catch (Exception e) {
-      Assertions.fail(e);
-    }
-    return this;
+  public TestAssert1Status assertStatus() {
+    return new TestAssertStatusImpl(this.actions, this.objectMapper);
   }
 
   /**
@@ -120,5 +86,21 @@ public final class TestAssertImpl implements TestAssert {
   @Override
   public TestAssertHead assertHead() {
     return new TestAssertHeadImpl(this.actions);
+  }
+
+  /**
+   * Asserts that the HTTP response matches custom validation logic.
+   *
+   * <p>This method returns an instance of {@code TestAssertCustom} for asserting custom validations
+   * on the HTTP response. It allows users to define their own result matchers or custom logic for
+   * validating the response, giving flexibility beyond standard status, content, and header
+   * assertions.
+   *
+   * @return an instance of {@code TestAssertCustom} for custom assertions on the response
+   * @since 1.1.0
+   */
+  @Override
+  public TestAssertCustom assertCustom() {
+    return new TestAssertCustomImpl(this.actions);
   }
 }