+++ src/test/java/io/github/co_mmer/aaamockmvc/test/testdata/testutil/TestObjectDto.java
@@ -1,3 +1,0 @@
-package io.github.co_mmer.aaamockmvc.test.testdata.testutil;
-
-public record TestObjectDto(int id, String name) {}